Some refactoring and tidying
